--- concert-reservation-application/src/main/java/concert/application/member/presentation/request/MemberRequest.java
@@ -1,17 +1,0 @@
-package concert.application.member.presentation.request;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class MemberRequest {
-
-    private String name;
-
-    @Builder
-    public MemberRequest(String name){
-        this.name = name;
-    }
-}
