--- concert-reservation-application/src/main/java/concert/application/member/presentation/response/MemberBalanceResponse.java
@@ -1,20 +1,0 @@
-package concert.application.member.presentation.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class MemberBalanceResponse {
-    private final long balance;
-
-    @Builder
-    public MemberBalanceResponse(long balance){
-        this.balance = balance;
-    }
-
-    public static MemberBalanceResponse of(long balance){
-        return MemberBalanceResponse.builder()
-                .balance(balance)
-                .build();
-    }
-}
