+++ concert-reservation-application/src/main/java/concert/application/waitingqueue/business/WaitingQueueFacade.java
@@ -1,9 +1,9 @@
 package concert.application.waitingqueue.business;
 
-import concert.application.waitingqueue.presentation.response.TokenResponse;
 import concert.domain.member.services.MemberService;
+import concert.domain.waitingqueue.entities.vo.TokenVO;
 import concert.domain.waitingqueue.services.WaitingQueueService;
-import concert.domain.waitingqueue.entities.vo.WaitingRankVo;
+import concert.domain.waitingqueue.entities.vo.WaitingRankVO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -14,14 +14,13 @@ public class WaitingQueueFacade {
   private final MemberService memberService;
   private final WaitingQueueService waitingQueueService;
 
-  public TokenResponse retrieveToken(long concertId, String uuid) {
+  public TokenVO retrieveToken(long concertId, String uuid) {
     memberService.getMemberByUuid(uuid);
-
     String token = waitingQueueService.retrieveToken(concertId, uuid);
-    return TokenResponse.of(token);
+    return TokenVO.of(token);
   }
 
-  public WaitingRankVo retrieveWaitingRank(long concertId, String uuid) {
+  public WaitingRankVO retrieveWaitingRank(long concertId, String uuid) {
     return waitingQueueService.retrieveWaitingRank(concertId, uuid);
   }
 }
