--- concert-reservation-application/src/main/java/concert/application/waitingqueue/presentation/response/WaitingRankResponse.java
@@ -1,24 +1,0 @@
-package concert.application.waitingqueue.presentation.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class WaitingRankResponse {
-
-    private final long waitingRank;
-    private final String status;
-
-    @Builder
-    public WaitingRankResponse(long waitingRank, String status){
-        this.waitingRank = waitingRank;
-        this.status = status;
-    }
-
-    public static WaitingRankResponse of(long waitingRank, String status){
-        return WaitingRankResponse.builder()
-                .waitingRank(waitingRank)
-                .status(status)
-                .build();
-    }
-}
