+++ concert-reservation-application/src/test/java/concert/concertscheduleseat/ConcertScheduleEntitySeatConcurrencyIntegrationTest.java
@@ -1,6 +1,6 @@
 package concert.concertscheduleseat;
 
-import concert.application.concertscheduleseat.business.ConcertScheduleSeatFacade;
+import concert.application.concert.business.ConcertScheduleSeatFacade;
 import concert.domain.concert.entities.ConcertEntity;
 import concert.domain.concert.entities.ConcertScheduleEntity;
 import concert.domain.concert.entities.dao.ConcertEntityDAO;
