--- concert-reservation-domain/src/main/java/concert/domain/waitingqueue/entities/vo/TokenResponse.java
@@ -1,20 +1,0 @@
-package concert.domain.waitingqueue.entities.vo;
-
-import lombok.Builder;
-import lombok.Getter;
-@Getter
-public class TokenResponse {
-
-    private final String token;
-
-    @Builder
-    public TokenResponse(String token){
-        this.token = token;
-    }
-
-    public static TokenResponse of(String token){
-        return TokenResponse.builder()
-                .token(token)
-                .build();
-    }
-}
