+++ concert-reservation-interfaces/src/main/java/concert/interfaces/concert/request/ConcertScheduleCreateRequest.java
@@ -1,0 +1,21 @@
+package main.java.concert.interfaces.concert.request;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Getter
+@NoArgsConstructor
+public class ConcertScheduleCreateRequest {
+
+    private String concertName;
+    private LocalDateTime dateTime;
+
+    @Builder
+    public ConcertScheduleCreateRequest(String concertName, LocalDateTime dateTime){
+        this.concertName = concertName;
+        this.dateTime = dateTime;
+    }
+}
