+++ concert-reservation-interfaces/src/main/java/concert/interfaces/concert/request/ConcertScheduleRequest.java
@@ -1,0 +1,19 @@
+package main.java.concert.interfaces.concert.request;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class ConcertScheduleRequest {
+
+    private String token;
+    private long concertId;
+
+    @Builder
+    public ConcertScheduleRequest(String token, long concertId){
+        this.token = token;
+        this.concertId = concertId;
+    }
+}
