+++ concert-reservation-interfaces/src/main/java/concert/interfaces/concert/request/SeatNumbersRequest.java
@@ -1,0 +1,18 @@
+package main.java.concert.interfaces.concert.request;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class SeatNumbersRequest {
+
+    private String token;
+    private long concertScheduleId;
+
+    public SeatNumbersRequest(String token, long concertScheduleId){
+        this.token = token;
+        this.concertScheduleId = concertScheduleId;
+    }
+
+}
