+++ concert-reservation-interfaces/src/main/java/concert/interfaces/member/request/MemberRequest.java
@@ -1,0 +1,17 @@
+package concert.interfaces.member.request;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class MemberRequest {
+
+    private String name;
+
+    @Builder
+    public MemberRequest(String name){
+        this.name = name;
+    }
+}
