+++ concert-reservation-interfaces/src/main/java/concert/interfaces/member/response/MemberBalanceResponse.java
@@ -1,0 +1,20 @@
+package concert.interfaces.member.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+public class MemberBalanceResponse {
+    private final long balance;
+
+    @Builder
+    public MemberBalanceResponse(long balance){
+        this.balance = balance;
+    }
+
+    public static MemberBalanceResponse of(long balance){
+        return MemberBalanceResponse.builder()
+                .balance(balance)
+                .build();
+    }
+}
