+++ concert-reservation-interfaces/src/main/java/concert/interfaces/waitingqueue/response/WaitingRankResponse.java
@@ -1,0 +1,24 @@
+package concert.interfaces.waitingqueue.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+public class WaitingRankResponse {
+
+    private final long waitingRank;
+    private final String status;
+
+    @Builder
+    public WaitingRankResponse(long waitingRank, String status){
+        this.waitingRank = waitingRank;
+        this.status = status;
+    }
+
+    public static WaitingRankResponse of(long waitingRank, String status){
+        return WaitingRankResponse.builder()
+                .waitingRank(waitingRank)
+                .status(status)
+                .build();
+    }
+}
