+++ settings.gradle
@@ -1,6 +1,7 @@
 rootProject.name = 'concert'
 
 include 'commons'
+include 'concert-reservation-interfaces'
 include 'concert-reservation-application'
 include 'concert-reservation-domain'
 include 'concert-reservation-infrastructure'
